chore: add more help texts in backup forms
