+++ ui/apps/platform/src/Containers/Integrations/IntegrationForm/Forms/ExternalBackupIntegrations/S3IntegrationForm.tsx
@@ -1,47 +1,57 @@
 /* eslint-disable no-void */
 import React, { ReactElement } from 'react';
-import {
-    Checkbox,
-    Form,
-    FormSelect,
-    PageSection,
-    TextInput,
-    TextArea,
-} from '@patternfly/react-core';
+import { Checkbox, Form, FormSelect, PageSection, TextInput } from '@patternfly/react-core';
 import * as yup from 'yup';
 
 import { BackupIntegrationBase } from 'services/BackupIntegrationsService';
 
 import usePageState from 'Containers/Integrations/hooks/usePageState';
 import FormMessage from 'Components/PatternFly/FormMessage';
+import FormCancelButton from 'Components/PatternFly/FormCancelButton';
 import FormTestButton from 'Components/PatternFly/FormTestButton';
 import FormSaveButton from 'Components/PatternFly/FormSaveButton';
-import FormCancelButton from 'Components/PatternFly/FormCancelButton';
-import useIntegrationForm from '../useIntegrationForm';
-import { IntegrationFormProps } from '../integrationFormTypes';
+import useIntegrationForm from '../../useIntegrationForm';
+import { IntegrationFormProps } from '../../integrationFormTypes';
+import { s3IamRole, s3EndpointIcon, objectPrefixIcon } from './icons';
 
-import IntegrationFormActions from '../IntegrationFormActions';
-import FormLabelGroup from '../FormLabelGroup';
-import ScheduleIntervalOptions from '../FormSchedule/ScheduleIntervalOptions';
-import ScheduleWeeklyOptions from '../FormSchedule/ScheduleWeeklyOptions';
-import ScheduleDailyOptions from '../FormSchedule/ScheduleDailyOptions';
+import IntegrationFormActions from '../../IntegrationFormActions';
+import FormLabelGroup from '../../FormLabelGroup';
+import ScheduleIntervalOptions from '../../FormSchedule/ScheduleIntervalOptions';
+import ScheduleWeeklyOptions from '../../FormSchedule/ScheduleWeeklyOptions';
+import ScheduleDailyOptions from '../../FormSchedule/ScheduleDailyOptions';
 
-import { getGoogleCredentialsPlaceholder } from '../../utils/integrationUtils';
-
-export type GcsIntegration = {
-    gcs: {
+export type S3Integration = {
+    s3: {
         bucket: string;
         objectPrefix: string;
-        useWorkloadId: boolean;
-        serviceAccount: string;
+        endpoint: string;
+        region: string;
+        useIam: boolean;
+        accessKeyId: string;
+        secretAccessKey: string;
     };
-    type: 'gcs';
+    type: 's3';
 } & BackupIntegrationBase;
 
-export type GcsIntegrationFormValues = {
-    externalBackup: GcsIntegration;
+export type S3IntegrationFormValues = {
+    externalBackup: S3Integration;
     updatePassword: boolean;
 };
+
+function requireCredentials(value, context: yup.TestContext) {
+    const requirePasswordField =
+        // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+        // @ts-ignore
+        context?.from[2]?.value?.updatePassword || false;
+    const useIam = context?.parent?.useIam;
+
+    if (!requirePasswordField || useIam) {
+        return true;
+    }
+
+    const trimmedValue = value?.trim();
+    return !!trimmedValue;
+}
 
 export const validationSchema = yup.object().shape({
     externalBackup: yup.object().shape({
@@ -58,42 +68,31 @@ export const validationSchema = yup.object().shape({
             hour: yup.number(),
             minute: yup.number(),
         }),
-        gcs: yup.object().shape({
+        s3: yup.object().shape({
             bucket: yup.string().trim().required('Bucket is required'),
-            objectPrefix: yup.string().trim(),
-            useWorkloadId: yup.bool(),
-            serviceAccount: yup
+            objectPrefix: yup.string(),
+            endpoint: yup.string(),
+            region: yup.string().trim().required('Region is required'),
+            useIam: yup.bool(),
+            accessKeyId: yup
+                .string()
+                .trim()
+                .test('accessKeyId-test', 'An access key ID is required', requireCredentials),
+            secretAccessKey: yup
                 .string()
                 .trim()
                 .test(
-                    'serviceAccount-test',
-                    'Valid JSON is required for service account key',
-                    (value, context: yup.TestContext) => {
-                        const requirePasswordField =
-                            // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-                            // @ts-ignore
-                            context?.from[2]?.value?.updatePassword || false;
-                        const useWorkloadId = context?.parent?.useWorkloadId;
-
-                        if (!requirePasswordField || useWorkloadId) {
-                            return true;
-                        }
-                        try {
-                            JSON.parse(value as string);
-                        } catch (e) {
-                            return false;
-                        }
-                        const trimmedValue = value?.trim();
-                        return !!trimmedValue;
-                    }
+                    'secretAccessKey-test',
+                    'A secret access key is required',
+                    requireCredentials
                 ),
         }),
-        type: yup.string().matches(/gcs/),
+        type: yup.string().matches(/s3/),
     }),
     updatePassword: yup.bool(),
 });
 
-export const defaultValues: GcsIntegrationFormValues = {
+export const defaultValues: S3IntegrationFormValues = {
     externalBackup: {
         id: '',
         name: '',
@@ -103,22 +102,26 @@ export const defaultValues: GcsIntegrationFormValues = {
             hour: 0,
             minute: 0,
         },
-        gcs: {
+        s3: {
             bucket: '',
             objectPrefix: '',
-            useWorkloadId: false,
-            serviceAccount: '',
+            endpoint: '',
+            region: '',
+            useIam: false,
+            accessKeyId: '',
+            secretAccessKey: '',
         },
-        type: 'gcs',
+        type: 's3',
     },
     updatePassword: true,
 };
 
-function GcsIntegrationForm({
+function S3IntegrationForm({
     initialValues = null,
     isEditable = false,
-}: IntegrationFormProps<GcsIntegration>): ReactElement {
+}: IntegrationFormProps<S3Integration>): ReactElement {
     const formInitialValues = { ...defaultValues, ...initialValues };
+
     if (initialValues) {
         formInitialValues.externalBackup = {
             ...formInitialValues.externalBackup,
@@ -126,7 +129,8 @@ function GcsIntegrationForm({
         };
         // We want to clear the password because backend returns '******' to represent that there
         // are currently stored credentials
-        formInitialValues.externalBackup.gcs.serviceAccount = '';
+        formInitialValues.externalBackup.s3.accessKeyId = '';
+        formInitialValues.externalBackup.s3.secretAccessKey = '';
 
         // Don't assume user wants to change password; that has caused confusing UX.
         formInitialValues.updatePassword = false;
@@ -145,7 +149,7 @@ function GcsIntegrationForm({
         onTest,
         onCancel,
         message,
-    } = useIntegrationForm<GcsIntegrationFormValues>({
+    } = useIntegrationForm<S3IntegrationFormValues>({
         initialValues: formInitialValues,
         validationSchema,
     });
@@ -155,15 +159,17 @@ function GcsIntegrationForm({
         return setFieldValue(event.target.id, value, false);
     }
 
-    function updateServiceAccountOnChange(value, event) {
+    function updateKeysOnChange(value, event) {
         void setFieldValue(event.target.id, value);
         if (value === true) {
-            void setFieldValue('externalBackup.gcs.serviceAccount', '');
+            void setFieldValue('externalBackup.s3.accessKeyId', '');
+            void setFieldValue('externalBackup.s3.secretAccessKey', '');
         }
     }
 
     function onUpdateCredentialsChange(value, event) {
-        setFieldValue('externalBackup.gcs.serviceAccount', '');
+        setFieldValue('externalBackup.s3.accessKeyId', '');
+        setFieldValue('externalBackup.s3.secretAccessKey', '');
         return setFieldValue(event.target.id, value);
     }
 
@@ -200,7 +206,6 @@ function GcsIntegrationForm({
                             isRequired
                             type="number"
                             id="externalBackup.backupsToKeep"
-                            name="externalBackup.backupsToKeep"
                             value={values.externalBackup.backupsToKeep}
                             onChange={(event, value) => onChange(value, event)}
                             onBlur={handleBlur}
@@ -263,16 +268,15 @@ function GcsIntegrationForm({
                     <FormLabelGroup
                         isRequired
                         label="Bucket"
-                        fieldId="externalBackup.gcs.bucket"
+                        fieldId="externalBackup.s3.bucket"
                         touched={touched}
                         errors={errors}
-                        helperText="example, stackrox.backups"
+                        helperText="example, acs.backups"
                     >
                         <TextInput
                             type="text"
-                            id="externalBackup.gcs.bucket"
-                            name="externalBackup.gcs.bucket"
-                            value={values.externalBackup.gcs.bucket}
+                            id="externalBackup.s3.bucket"
+                            value={values.externalBackup.s3.bucket}
                             onChange={(event, value) => onChange(value, event)}
                             onBlur={handleBlur}
                             isDisabled={!isEditable}
@@ -280,15 +284,49 @@ function GcsIntegrationForm({
                     </FormLabelGroup>
                     <FormLabelGroup
                         label="Object prefix"
-                        fieldId="externalBackup.gcs.objectPrefix"
+                        labelIcon={objectPrefixIcon()}
+                        fieldId="externalBackup.s3.objectPrefix"
                         touched={touched}
                         errors={errors}
                     >
                         <TextInput
                             type="text"
-                            id="externalBackup.gcs.objectPrefix"
-                            name="externalBackup.gcs.objectPrefix"
-                            value={values.externalBackup.gcs.objectPrefix}
+                            id="externalBackup.s3.objectPrefix"
+                            value={values.externalBackup.s3.objectPrefix}
+                            onChange={(event, value) => onChange(value, event)}
+                            onBlur={handleBlur}
+                            isDisabled={!isEditable}
+                        />
+                    </FormLabelGroup>
+                    <FormLabelGroup
+                        label="Endpoint"
+                        labelIcon={s3EndpointIcon()}
+                        fieldId="externalBackup.s3.endpoint"
+                        helperText="example, s3.us-west-2.amazonaws.com"
+                        touched={touched}
+                        errors={errors}
+                    >
+                        <TextInput
+                            type="text"
+                            id="externalBackup.s3.endpoint"
+                            value={values.externalBackup.s3.endpoint}
+                            onChange={(event, value) => onChange(value, event)}
+                            onBlur={handleBlur}
+                            isDisabled={!isEditable}
+                        />
+                    </FormLabelGroup>
+                    <FormLabelGroup
+                        isRequired
+                        label="Region"
+                        fieldId="externalBackup.s3.region"
+                        helperText="example, us-west-2"
+                        touched={touched}
+                        errors={errors}
+                    >
+                        <TextInput
+                            type="text"
+                            id="externalBackup.s3.region"
+                            value={values.externalBackup.s3.region}
                             onChange={(event, value) => onChange(value, event)}
                             onBlur={handleBlur}
                             isDisabled={!isEditable}
@@ -296,15 +334,16 @@ function GcsIntegrationForm({
                     </FormLabelGroup>
                     <FormLabelGroup
                         label=""
-                        fieldId="externalBackup.gcs.useWorkloadId"
+                        labelIcon={s3IamRole()}
+                        fieldId="externalBackup.s3.useIam"
                         touched={touched}
                         errors={errors}
                     >
                         <Checkbox
-                            label="Use workload identity"
-                            id="externalBackup.gcs.useWorkloadId"
-                            isChecked={values.externalBackup.gcs.useWorkloadId}
-                            onChange={(event, value) => updateServiceAccountOnChange(value, event)}
+                            label="Use container IAM role"
+                            id="externalBackup.s3.useIam"
+                            isChecked={values.externalBackup.s3.useIam}
+                            onChange={(event, value) => updateKeysOnChange(value, event)}
                             onBlur={handleBlur}
                             isDisabled={!isEditable}
                         />
@@ -314,51 +353,68 @@ function GcsIntegrationForm({
                             label=""
                             fieldId="updatePassword"
                             helperText="Enable this option to replace currently stored credentials (if any)"
-                            touched={touched}
                             errors={errors}
                         >
                             <Checkbox
-                                label="Update stored credentials"
+                                label="Update access key ID and secret access key"
                                 id="updatePassword"
-                                isChecked={
-                                    !values.externalBackup.gcs.useWorkloadId &&
-                                    values.updatePassword
-                                }
+                                isChecked={values.updatePassword}
                                 onChange={(event, value) => onUpdateCredentialsChange(value, event)}
                                 onBlur={handleBlur}
-                                isDisabled={!isEditable || values.externalBackup.gcs.useWorkloadId}
+                                isDisabled={!isEditable}
                             />
                         </FormLabelGroup>
                     )}
                     <FormLabelGroup
-                        label="Service account key (JSON)"
-                        isRequired={
-                            values.updatePassword && !values.externalBackup.gcs.useWorkloadId
-                        }
-                        fieldId="externalBackup.gcs.serviceAccount"
+                        label="Access key ID"
+                        fieldId="externalBackup.s3.accessKeyId"
+                        isRequired={values.updatePassword && !values.externalBackup.s3.useIam}
                         touched={touched}
                         errors={errors}
                     >
-                        <TextArea
-                            className="json-input"
-                            isRequired={
-                                values.updatePassword && !values.externalBackup.gcs.useWorkloadId
-                            }
-                            type="text"
-                            id="externalBackup.gcs.serviceAccount"
-                            name="externalBackup.gcs.serviceAccount"
-                            value={values.externalBackup.gcs.serviceAccount}
+                        <TextInput
+                            isRequired={values.updatePassword && !values.externalBackup.s3.useIam}
+                            type="password"
+                            id="externalBackup.s3.accessKeyId"
+                            value={values.externalBackup.s3.accessKeyId}
                             onChange={(event, value) => onChange(value, event)}
                             onBlur={handleBlur}
                             isDisabled={
                                 !isEditable ||
                                 !values.updatePassword ||
-                                values.externalBackup.gcs.useWorkloadId
+                                values.externalBackup.s3.useIam
                             }
-                            placeholder={getGoogleCredentialsPlaceholder(
-                                values.externalBackup.gcs.useWorkloadId,
-                                values.updatePassword
-                            )}
+                            placeholder={
+                                values.updatePassword || values.externalBackup.s3.useIam
+                                    ? ''
+                                    : 'Currently-stored access key ID will be used.'
+                            }
+                        />
+                    </FormLabelGroup>
+                    <FormLabelGroup
+                        label="Secret access key"
+                        fieldId="externalBackup.s3.secretAccessKey"
+                        isRequired={values.updatePassword && !values.externalBackup.s3.useIam}
+                        touched={touched}
+                        errors={errors}
+                    >
+                        <TextInput
+                            isRequired={values.updatePassword && !values.externalBackup.s3.useIam}
+                            type="password"
+                            id="externalBackup.s3.secretAccessKey"
+                            value={values.externalBackup.s3.secretAccessKey}
+                            onChange={(event, value) => onChange(value, event)}
+                            onBlur={handleBlur}
+                            isDisabled={
+                                !isEditable ||
+                                !values.updatePassword ||
+                                values.externalBackup.s3.useIam
+                            }
+                            placeholder={
+                                values.updatePassword || values.externalBackup.s3.useIam
+                                    ? ''
+                                    : 'Currently-stored secret access key will be used.'
+                            }
                         />
                     </FormLabelGroup>
                 </Form>
@@ -388,4 +444,4 @@ function GcsIntegrationForm({
     );
 }
 
-export default GcsIntegrationForm;
+export default S3IntegrationForm;
